refactor: ideation -> response collection
