+++ src/modules/main/PlayerView.tsx
@@ -1,6 +1,6 @@
 import { PLAYER_VIEW_CY } from '@/config/selectors';
 
-import Ideation from '../inputResponse/Ideation';
+import Ideation from '../responseCollection/ResponseCollection';
 
 const PlayerView = (): JSX.Element => (
   <div data-cy={PLAYER_VIEW_CY}>
